Style navbar and footer
