+++ client/src/components/Game/Game.jsx
@@ -1,7 +1,16 @@
-import React from 'react';
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import * as actions from '../../actions';
+import axios from 'axios';
 
-class Game extends React.Component {
+class Game extends Component {
+  constructor(props) {
+    super(props);
+  }
 
+  render() {
+    return <div></div>;
+  }
 }
 
-export default Game;
+export default connect(null, actions)(Game);
