--- client/src/components/Lobby/Lobby.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import * as actions from '../../actions';
-import axios from 'axios';
-
-class Lobby extends Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return <div></div>;
-  }
-}
-
-export default connect(null, actions)(Lobby);
