+++ client/src/components/NavBar/NavBar.jsx
@@ -3,8 +3,6 @@ import { connect } from 'react-redux';
 import * as actions from '../../actions';
 import axios from 'axios';
 
-import './NavBar.css';
-
 class NavBar extends Component {
   constructor(props) {
     super(props);
