+++ client/src/components/Session/Session.jsx
@@ -1,7 +1,16 @@
-import React from 'react';
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import * as actions from '../../actions';
+import axios from 'axios';
 
-class Session extends React.Component {
+class Session extends Component {
+  constructor(props) {
+    super(props);
+  }
 
+  render() {
+    return <div></div>;
+  }
 }
 
-export default Session;
+export default connect(null, actions)(Session);
