+++ client/src/index.js
@@ -2,8 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Routes from './routes';
 
-import './reset.css';
-import './index.css';
+import './styles/reset.css';
+import './styles/index.css';
+import './styles/NavBar.css';
+import './styles/Footer.css';
 
 import { Provider } from 'react-redux';
 import { applyMiddleware, createStore } from 'redux';
