+++ client/src/styles/Footer.css
@@ -1,0 +1,24 @@
+footer {
+  width: 100%;
+  padding: 10px 0;
+  background: #161925;
+  color: #dfdfdf;
+}
+
+.footer-inner {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 80%;
+  margin: 0 auto;
+}
+
+.footer-links a {
+  margin-left: 20px;
+  color: #8EA8C3;
+  transition: all 0.3s ease;
+}
+
+.footer-links a:hover {
+  color: #CBF7ED;
+}
