+++ client/src/styles/NavBar.css
@@ -1,0 +1,50 @@
+.navbar {
+  width: 100%;
+  padding: 10px 0;
+  background: #23395B;
+  color: #CBF7ED;
+}
+
+.inside-navbar {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  position: relative;
+  margin: 0 auto;
+  width: 80%;
+}
+
+.inside-navbar input {
+  margin-left: 10px;
+  padding: 0 5px;
+  border: 1px solid #406E8E;
+  border-radius: 3px;
+  transition: all 0.3s ease;
+}
+
+.inside-navbar button {
+  margin-left: 10px;
+  background: transparent;
+  border: 2px solid #8EA8C3;
+  border-radius: 3px;
+  font-size: 14px;
+  text-transform: uppercase;
+  color: #8EA8C3;
+  transition: all 0.3s ease;
+}
+
+.inside-navbar button:hover {
+  border-color: #CBF7ED;
+  background: #CBF7ED;
+  color: #23395B;
+}
+
+.navbar-error {
+  position: absolute;
+  font-size: 12px;
+  font-weight: bold;
+  text-transform: uppercase;
+  color: #fff;
+  bottom: -10px;
+  right: 0;
+}
