+++ client/src/styles/Splash.css
@@ -1,0 +1,7 @@
+.splash-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  flex: 1;
+}
