+++ client/src/styles/index.css
@@ -1,0 +1,7 @@
+body {
+  margin: 0;
+  padding: 0;
+  font-family: Heebo, Lato;
+  background: #406E8E;
+  color: #161925;
+}
